Task logic working.

+++ src/todos/todos.controller.ts
@@ -7,7 +7,7 @@ import {
   Param,
   Delete,
   UseGuards,
-  Req,
+  Request,
   UseInterceptors,
   UploadedFile,
   Res,
@@ -28,8 +28,8 @@ export class TodosController {
   constructor(private readonly todosService: TodosService) {}
 
   @Post()
-  create(@Body() createTodoDto: CreateTodoDto, @Req() req) {
-    createTodoDto.creator = req.user._id;
+  create(@Body() createTodoDto: CreateTodoDto, @Request() req) {
+    createTodoDto.creator = req.user.id;
     return this.todosService.create(createTodoDto);
   }
 
@@ -39,18 +39,86 @@ export class TodosController {
   }
 
   @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.todosService.findOne(id);
+  async findOne(@Param('id') id: string, @Request() req, @Res() res) {
+    try {
+      const task = await this.todosService.findOne(id);
+      if (!task) {
+        return res.status(HttpStatus.NOT_FOUND).json({
+          msg: 'No task found.',
+        });
+      }
+      const creator = (task.creator as any)._id.toString();
+
+      if (task.private && creator != req.user.id) {
+        return res.status(HttpStatus.UNAUTHORIZED).json({
+          msg: 'You are not allowed to view private tasks.',
+        });
+      }
+      return res.status(HttpStatus.OK).json(task);
+    } catch (e) {
+      return res.status(HttpStatus.NOT_FOUND).json({
+        msg: 'No task found.',
+      });
+    }
   }
 
   @Put(':id')
-  update(@Param('id') id: string, @Body() updateTodoDto: UpdateTodoDto) {
-    return this.todosService.update(id, updateTodoDto);
+  async update(
+    @Param('id') id: string,
+    @Body() updateTodoDto: UpdateTodoDto,
+    @Request() req,
+    @Res() res,
+  ) {
+    try {
+      const task = await this.todosService.findOne(id);
+      if (!task) {
+        return res.status(HttpStatus.NOT_FOUND).json({
+          msg: 'No task found.',
+        });
+      }
+      const creator = (task.creator as any)._id.toString();
+
+      if (task.private && creator != req.user.id) {
+        return res.status(HttpStatus.UNAUTHORIZED).json({
+          msg: 'You are not allowed to update private tasks of oither people.',
+        });
+      }
+
+      const todo = await this.todosService.update(id, updateTodoDto);
+      return res.status(HttpStatus.OK).json(todo);
+    } catch (e) {
+      return res.status(HttpStatus.NOT_FOUND).json({
+        msg: 'No task found.',
+      });
+    }
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.todosService.remove(id);
+  async remove(@Param('id') id: string, @Request() req, @Res() res) {
+    try {
+      const task = await this.todosService.findOne(id);
+      if (!task) {
+        return res.status(HttpStatus.NOT_FOUND).json({
+          msg: 'No task found.',
+        });
+      }
+      console.log('the task: ', task);
+
+      const creator = (task.creator as any)._id.toString();
+
+      if (task.private && creator != req.user.id) {
+        return res.status(HttpStatus.UNAUTHORIZED).json({
+          msg: 'You are not allowed to delete private tasks of oither people.',
+        });
+      }
+
+      await this.todosService.remove(id);
+      return res.status(HttpStatus.OK).json({ msg: 'Task deleted.' });
+    } catch (e) {
+      return res.status(HttpStatus.NOT_FOUND).json({
+        msg: 'No task found.',
+      });
+    }
   }
 
   @Post(':id/img')
@@ -71,7 +139,7 @@ export class TodosController {
   async uploadFile(
     @Param('id') todoId,
     @UploadedFile() file,
-    @Req() req,
+    @Request() req,
     @Res() res,
   ) {
     const url = req.protocol + '://' + req.get('host') + '/todos/' + file.path;
